refactor(ui): Move edit_item to common/edit_item
