+++ apps/clearance_ui/src/components/clearance_inspector/LicenseConclusionItem.tsx
@@ -15,9 +15,9 @@ import {
     TooltipProvider,
     TooltipTrigger,
 } from "@/components/ui/tooltip";
+import EditButton from "@/components/common/edit_item/EditButton";
 import PurlDetails from "@/components/common/PurlDetails";
 import DeleteLicenseConclusion from "@/components/delete_item/DeleteLicenseConclusion";
-import EditButton from "@/components/edit_item/EditButton";
 import { stringToColourRGBA } from "@/helpers/stringToColour";
 import { cn } from "@/lib/utils";
