+++ apps/clearance_ui/src/components/main_ui/package_bulk_conclusions/BulkConclusion.tsx
@@ -8,8 +8,8 @@ import { userAPI } from "validation-helpers";
 import { userHooks } from "@/hooks/zodiosHooks";
 import { Badge } from "@/components/ui/badge";
 import { Separator } from "@/components/ui/separator";
+import EditButton from "@/components/common/edit_item/EditButton";
 import DeleteBulkConclusion from "@/components/delete_item/DeleteBulkConclusion";
-import EditButton from "@/components/edit_item/EditButton";
 
 type BulkConclusion = ZodiosResponseByAlias<
     typeof userAPI,
