+++ apps/clearance_ui/src/components/path_exclusions/ExclusionDB.tsx
@@ -19,8 +19,8 @@ import {
     PopoverContent,
     PopoverTrigger,
 } from "@/components/ui/popover";
+import EditButton from "@/components/common/edit_item/EditButton";
 import DeletePathExclusion from "@/components/delete_item/DeletePathExclusion";
-import EditButton from "@/components/edit_item/EditButton";
 import ExclusionFormDialog from "@/components/path_exclusions/ExclusionFormDialog";
 import { toPathPurl } from "@/helpers/pathParamHelpers";
